# Conflicts:
#   Trigger/TrigMonitoring/TrigEgammaMonitoring/python/TrigEgammaMonitorAlgorithm.py
#   Trigger/TrigMonitoring/TrigEgammaMonitoring/src/TrigEgammaMonitorAlgorithm.cxx
#   Trigger/TrigMonitoring/TrigEgammaMonitoring/src/TrigEgammaMonitorAlgorithm.h

+++ MuonSpectrometer/MuonCnv/MuonEventTPCnv/MuonEventTPCnv/MuonDigitContainer/MuonDigitContainer_p2.h
@@ -34,8 +34,9 @@ class MuonDigitContainer_p2
   std::vector<uint16_t>  m_size;
   
   std::vector<DIGIT>                           m_digits;
-   /** The delta identifiers of the PRD i.e. we only store the difference from the last one. This saves space, compared to always writing the full 32-bit word*/
-  std::vector<unsigned short>                  m_digitDeltaId;
+
+  /** The 32-bits compact identifier */
+  std::vector<unsigned int>                  m_digitDeltaId;
 };
 
 typedef MuonDigitContainer_p2<MM_Digit_p2> MM_DigitContainer_p2;
